+++ projects/mind-mapp/app/src/utils/index.ts
@@ -10,6 +10,7 @@ export * from './fitViewMath';
 export * from './focusHistory';
 export * from './focusHistoryStorage';
 export * from './focusNav';
+export * from './helpStatus';
 export * from './helpToggle';
 export * from './id';
 export * from './importValidation';